Modify isEdgeTouched method to work with full screen gestures
In order to make full screen swipe back possible I had to modify this function to return true only when the drag has started instead of when edge was touched.

This change was required as with full screen swipe back there is no such thing as edges from which users can start the drag - the whole activity can be dragged instead.